renames, file moves, relnotes, build flags
