+++ source/common/event/dispatcher_impl.cc
@@ -26,7 +26,7 @@
 #include "event2/event.h"
 
 #ifdef ENVOY_HANDLE_SIGNALS
-#include "exe/signal_action.h"
+#include "common/signal/signal_action.h"
 #endif
 
 namespace Envoy {
