+++ source/common/signal/fatal_error_handler.h
