+++ source/common/signal/signal_action.h
@@ -7,8 +7,7 @@
 #include <list>
 
 #include "common/common/non_copyable.h"
-
-#include "exe/fatal_error_handler.h"
+#include "common/signal/fatal_error_handler.h"
 
 #include "server/backtrace.h"
