+++ source/exe/BUILD
@@ -31,11 +31,6 @@ envoy_cc_binary(
 )
 
 envoy_cc_library(
-    name = "fatal_error_handler_lib",
-    hdrs = ["fatal_error_handler.h"],
-)
-
-envoy_cc_library(
     name = "envoy_common_lib",
     deps = [
         "//source/common/event:libevent_lib",
@@ -81,7 +76,7 @@ envoy_cc_library(
     ] + select({
         "//bazel:disable_signal_trace": [],
         "//conditions:default": [
-            ":sigaction_lib",
+            "//source/common/signal:sigaction_lib",
             ":terminate_handler_lib",
         ],
     }) + envoy_cc_platform_dep("platform_impl_lib"),
@@ -118,20 +113,6 @@ envoy_cc_win32_library(
         "//source/common/common:assert_lib",
         "//source/common/common:thread_lib",
         "//source/common/filesystem:filesystem_lib",
-    ],
-)
-
-envoy_cc_library(
-    name = "sigaction_lib",
-    srcs = ["signal_action.cc"],
-    hdrs = ["signal_action.h"],
-    tags = ["backtrace"],
-    deps = [
-        ":fatal_error_handler_lib",
-        "//source/common/common:assert_lib",
-        "//source/common/common:non_copyable",
-        "//source/common/singleton:threadsafe_singleton",
-        "//source/server:backtrace_lib",
     ],
 )
