+++ test/exe/BUILD
@@ -53,16 +53,6 @@ envoy_cc_test(
 )
 
 envoy_cc_test(
-    name = "signals_test",
-    srcs = ["signals_test.cc"],
-    tags = ["backtrace"],
-    deps = [
-        "//source/exe:sigaction_lib",
-        "//test/test_common:utility_lib",
-    ],
-)
-
-envoy_cc_test(
     name = "terminate_handler_test",
     srcs = ["terminate_handler_test.cc"],
     tags = ["backtrace"],
